[refactor]: split http-api router by resource
router.rs (766 LOC) was a single file holding routes, handlers, and tests for
every endpoint. Pure layout split into router/{health,focuses,proposals,tests}.rs
with a thin mod.rs that owns AppState, ApiError, FocusCatalogEntry, and the
two router builders.

Each resource module exposes a Router<AppState> via routes(); mod.rs merges
them. No logic moves; tests still exercise the assembled router end-to-end.
